Support sharpbang in dabsic file. Only values between [0;1] for analog input and output of VM110n are used now. Some land prepared for multiformat pixelarray and multiformat sound effect. Remove cinematic context (soon a programmable clipable context will be commited)

+++ src/history/count_events.cpp
@@ -24,7 +24,7 @@ int64_t	bunny_history_count_events_of(t_bunny_event_history *_his,
   auto it = his->events.lower_bound(tick);
   int64_t	z = 0;
 
-  if (it->first != tick)
+  if ((int64_t)it->first != tick)
     return (0);
   z += it->second.size();
   return (z);
